backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "app",email = "app@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.9,<5.0.0)",
    "django (>=6.0.3,<7.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "astropy (>=6.1)",
    "healpy (>=1.17)",
    "requests (>=2.32)",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
